@@ -3,10 +3,13 @@ package com.dex.teufelsturmoffline.views;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+
 import android.util.Pair;
 import android.view.LayoutInflater;
 import android.view.Menu;
@@ -17,12 +20,10 @@ import android.view.ViewGroup;
 import android.widget.AdapterView;
 import android.widget.SearchView;
 import android.widget.Spinner;
-import android.widget.Toast;
 
 
 import com.dex.teufelsturmoffline.R;
 import com.dex.teufelsturmoffline.activities.CommentActivity;
-import com.dex.teufelsturmoffline.activities.MapActivity;
 import com.dex.teufelsturmoffline.adapter.RouteRecycleAdapter;
 import com.dex.teufelsturmoffline.adapter.SpinnerAreaAdapter;
 import com.dex.teufelsturmoffline.database.DatabaseHelper;
@@ -30,13 +31,11 @@ import com.dex.teufelsturmoffline.database.SettingsSaver;
 import com.dex.teufelsturmoffline.model.AreaSpinnerData;
 import com.dex.teufelsturmoffline.model.Route;
 
-import java.lang.reflect.Array;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
-public class SearchViewFragment extends Fragment implements View.OnClickListener {
+public class SearchViewFragment extends Fragment {
     Spinner spinner_area;
     RecyclerView recyclerView;
     RouteRecycleAdapter routeRecycleAdapter;
@@ -49,7 +48,8 @@ public class SearchViewFragment extends Fragment implements View.OnClickListener
     FloatingActionButton fab_map;
 
 
-    public SearchViewFragment(){}
+    public SearchViewFragment() {
+    }
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -59,25 +59,24 @@ public class SearchViewFragment extends Fragment implements View.OnClickListener
     }
 
 
-
     @Override
     public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
         inflater.inflate(R.menu.search_menu, menu);
-        MenuItem searchButton = menu.findItem( R.id.search);
+        MenuItem searchButton = menu.findItem(R.id.search);
         searchView = (SearchView) searchButton.getActionView();
         searchView.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
-
 
 
             @Override
             public boolean onQueryTextSubmit(String query) {
                 currentFilterText = query;
-                if(query.equals(""))
+                if (query.equals(""))
                     resetRouteList();
                 else
                     filterRouteList(query);
                 return false;
             }
+
             @Override
             public boolean onQueryTextChange(String s) {
                 currentFilterText = s;
@@ -89,10 +88,6 @@ public class SearchViewFragment extends Fragment implements View.OnClickListener
             }
         });
         super.onCreateOptionsMenu(menu, inflater);
-
-        fab_map = view.findViewById(R.id.fab_map);
-        fab_map.setOnClickListener(this);
-
     }
 
     @Override
@@ -105,8 +100,8 @@ public class SearchViewFragment extends Fragment implements View.OnClickListener
         List<Pair<String, Integer>> areaPairs = db.getAreas();
         List<AreaSpinnerData> spinnerData = new ArrayList<>();
 
-        for (Pair<String, Integer> pair : areaPairs){
-            spinnerData.add(new AreaSpinnerData(pair.first,pair.second));
+        for (Pair<String, Integer> pair : areaPairs) {
+            spinnerData.add(new AreaSpinnerData(pair.first, pair.second));
         }
 
         SpinnerAreaAdapter spinnerAreaAdapter = new SpinnerAreaAdapter(view.getContext(), R.layout.row_spinner_area, spinnerData);
@@ -127,7 +122,7 @@ public class SearchViewFragment extends Fragment implements View.OnClickListener
         spinner_area.setSelection(SettingsSaver.getArea(getActivity()));
 
         List<Route> routeList = new ArrayList<>();
-        if(spinnerData.size() > 0 ){
+        if (spinnerData.size() > 0) {
             routeList = db.getRoutesByArea(spinnerData.get(0).getArea());
         }
         LinearLayoutManager mLinearLayoutManager = new LinearLayoutManager(getActivity(),
@@ -145,20 +140,20 @@ public class SearchViewFragment extends Fragment implements View.OnClickListener
         recyclerView.setAdapter(routeRecycleAdapter);
 
 
-
         return view;
     }
 
-    public void updateRouteList(String area){
+    public void updateRouteList(String area) {
         List<Route> list = db.getRoutesByArea(area);
         Collections.sort(list);
         routeRecycleAdapter.updateData(list);
     }
-    public void filterRouteList(String filter){
+
+    public void filterRouteList(String filter) {
         routeRecycleAdapter.filterData(filter);
     }
 
-    public void resetRouteList(){
+    public void resetRouteList() {
         routeRecycleAdapter.resetData();
     }
 
@@ -166,17 +161,11 @@ public class SearchViewFragment extends Fragment implements View.OnClickListener
     public void onResume() {
         super.onResume();
         recyclerView.requestFocus();
-        if(db == null){
+        if (db == null) {
             db = new DatabaseHelper(getContext());
         }
         routeRecycleAdapter.updateData(db.getRoutesByArea(spinner_area.getSelectedItem().toString()));
         if (!currentFilterText.equals(""))
             filterRouteList(currentFilterText);
-    }
-
-    @Override
-    public void onClick(View view) {
-        Intent intent = new Intent(view.getContext(), MapActivity.class);
-        startActivity(intent);
     }
 }
